[GHA] fixes for trusty, xenial, eoan, and sid
For trusty and xenial, there are some packages that aren't pulled in by
buildessential such as gcc, automake, and libtool. So add those
explicitly. For eoan, we need to patch the sources.list file to get the
old release packages.

For sid, the compat version we are using is too old so bump it from 5 to
7.